Adds create_collection and load_index_time_boosting_dataframe to respective collection classes, simplifying index_time_boosts.py and spark/__init__.py Local engine memory limit set to 512m Solr create collection logging improvement Neatens indexing logic for multi-engine indexing Improves variable naming in aips/__init__.py Removes unnecessary code in weaviatecollection, also correctly deletes the index when overwrite is used in write()
