+++ aips/indexer.py
@@ -151,25 +151,27 @@ def build_collection(engine, dataset, force_rebuild=False, log=False):
     """
     source_datasets = dataset_info[dataset].get("source_datasets", [dataset])
     expected_count = sum([dataset_info[r]["count"] for r in source_datasets])
-    if force_rebuild or not engine.is_collection_healthy(dataset, expected_count, log=log):
-        if log: print(f"Reindexing [{dataset}] collection")
-        collection = engine.create_collection(dataset, log=log)
-        local_engine_collection = None
-        if not is_feature_supported(engine, dataset):
-            local_engine_collection = get_local_engine(log=log).create_collection(dataset, log=log)
-        overwrite = len(source_datasets) == 1
-        for dataset in source_datasets:
-            csv_file_path = download_data_files(dataset, log=log)
-            loader_args = dataset_info[dataset].get("loader_args", {})
-            dataframe = dataset_info[dataset]["loader_fn"](csv_file_path, **loader_args)
-            if dataset_info[dataset].get("enable_ltr", False):
-                get_ltr_engine(collection).enable_ltr()
-            collection.write(dataframe, overwrite=overwrite)
-            if local_engine_collection:
-                local_engine_collection.write(dataframe, overwrite=overwrite)
-    else:
-        if log: print(f"Collection [{dataset}] is healthy")
-        collection = engine.get_collection(dataset)
+    engines = [engine]
+    if not is_feature_supported(engine, dataset):
+        engines.append(get_local_engine(log=log))
+    for engine in engines:
+        if force_rebuild or not engine.is_collection_healthy(dataset, expected_count, log=log):
+            if log: print(f"Reindexing [{dataset}] collection")
+            collection = engine.create_collection(dataset, log=log)
+            local_engine_collection = None
+            overwrite = len(source_datasets) == 1
+            for dataset in source_datasets:
+                csv_file_path = download_data_files(dataset, log=log)
+                loader_args = dataset_info[dataset].get("loader_args", {})
+                dataframe = dataset_info[dataset]["loader_fn"](csv_file_path, **loader_args)
+                if dataset_info[dataset].get("enable_ltr", False):
+                    get_ltr_engine(collection).enable_ltr()
+                collection.write(dataframe, overwrite=overwrite)
+                if local_engine_collection:
+                    local_engine_collection.write(dataframe, overwrite=overwrite)
+        else:
+            if log: print(f"Collection [{dataset}] is healthy")
+            collection = engine.get_collection(dataset)
     return collection
 
 def copy_repository(dataset, log=False):
